backend done

+++ backend/controllers/product.controller.js
@@ -48,10 +48,13 @@ export const updaeProduct = async(req,res)=>{
 
 export const deleteProduct = async (req,res)=>{
     const {id} = req.params
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({success:false,message:"Product not found"})
+    }
     try {
         await Product.findByIdAndDelete(id)
         return res.status(200).json({success:true,message:"Product deleted"})
     } catch (error) {
-        return res.status(404).json({success:false,message:"Product not found"})
+        return res.status(500).json({success:false,message:"Server Error"})
     }
 }
